requirements.txt
aiohttp<3.10
aiosignal
aiosqlite
attrs
blinker
certifi
charset-normalizer
click
discord-protos
discord.py-self==2.1.0
Flask[async]
frozenlist
idna
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mdurl
multidict
numpy
pillow
playsound3
plyer
protobuf
psutil
Pygments
pytz
requests
rich
tzlocal
urllib3
Werkzeug
yarl
python-dotenv
pypresence